requirements.txt
requests==2.32.3
google-auth==2.29.0
google-api-python-client==2.131.0
